website: playground updates
